Add set_webhook management command
