+++ mpact_api/mpact/views.py
@@ -4,7 +4,8 @@ from rest_framework.views import APIView
 
 from .constants import MESSAGE, NEW_CHAT_PARTICIPANT, NEW_CHAT_TITLE, OK, TEXT
 from .logger import logger
-from .services import anonymize, set_webhook
+from .services import anonymize
+from .management.commands.set_webhook import set_webhook
 from .telegramevents import TelegramEvents
 
 
@@ -14,8 +15,8 @@ class Webhook(APIView):
     """
 
     def get(self, request):
-        webhook = set_webhook()
-        if webhook:
+        success = set_webhook()
+        if success:
             return Response("webhook setup ok")
         else:
             return Response("webhook setup failed")
